% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{useShinydashboard}
\alias{useShinydashboard}
\title{Warning: useShinydashboard has been deprecated and will be removed in a future
release of shinyWidgets. If you want to create value box in shiny, see package
bslib. If you absolutely need to use this function, copy the source code into
your project https://github.com/dreamRs/shinyWidgets/blob/26838f9e9ccdc90a47178b45318d110f5812d6e1/R/useShinydashboard.R}
\usage{
useShinydashboard()
}
\description{
Warning: useShinydashboard has been deprecated and will be removed in a future
release of shinyWidgets. If you want to create value box in shiny, see package
bslib. If you absolutely need to use this function, copy the source code into
your project https://github.com/dreamRs/shinyWidgets/blob/26838f9e9ccdc90a47178b45318d110f5812d6e1/R/useShinydashboard.R
}
